+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on Keep a Changelog (https://keepachangelog.com), and this p
 ## [Unreleased] - 2026-02-05
 ### Fixed
 - Fix misuse of assertions in `StripANSI` unit tests.
+- Add unit tests for `VarType` and `VarTypeAsInt`.
 
 ## [Unreleased] - 2026-02-04
 ### Added